refactor: make admin compile, scope project models to backends
