+++ .gitignore
@@ -17,7 +17,7 @@ recipe.json
 .idea/
 
 # Credentials for running ACME commands
-acme.json
+acme*.json
 
 # VS Code
 .vscode/
