+++ admin/src/args.rs
@@ -1,7 +1,6 @@
 use std::{fs, io, path::PathBuf};
 
 use clap::{Error, Parser, Subcommand};
-use shuttle_common::models::project::ProjectName;
 
 #[derive(Parser, Debug)]
 pub struct Args {
@@ -54,7 +53,7 @@ pub enum AcmeCommand {
 
         /// Project to request certificate for
         #[arg(long)]
-        project: ProjectName,
+        project: String,
 
         /// Path to acme credentials file
         /// This should have been created with `acme create-account`
@@ -70,7 +69,7 @@ pub enum AcmeCommand {
 
         /// Project to renew the certificate for
         #[arg(long)]
-        project: ProjectName,
+        project: String,
 
         /// Path to acme credentials file
         /// This should have been created with `acme create-account`
