+++ common/src/models/mod.rs
@@ -1,5 +1,7 @@
+pub mod admin;
 pub mod deployment;
 pub mod error;
+#[cfg(feature = "backend")]
 pub mod project;
 pub mod resource;
 pub mod secret;
